refactor: streamline receipt service imports and enhance receipt handling
- Updated import paths for receipt services to improve code organization.
- Removed user ID dependency in receipt fetching logic for broader data retrieval.
- Changed draft ID type from number to string in receipt-related hooks for consistency with API responses.
- Adjusted receipt creation and deletion logic to align with new API structure.
